predict StackSystem GetVerbsEvent and stack splits
