+++ Content.Server/Stack/StackSystem.cs
@@ -16,13 +16,9 @@ namespace Content.Server.Stack
     {
         [Dependency] private readonly IPrototypeManager _prototypeManager = default!;
 
-        public static readonly int[] DefaultSplitAmounts = { 1, 5, 10, 20, 30, 50 };
-
         public override void Initialize()
         {
             base.Initialize();
-
-            SubscribeLocalEvent<StackComponent, GetVerbsEvent<AlternativeVerb>>(OnStackAlternativeInteract);
         }
 
         public override void SetCount(EntityUid uid, int amount, StackComponent? component = null)
@@ -31,41 +27,16 @@ namespace Content.Server.Stack
                 return;
 
             base.SetCount(uid, amount, component);
-
-            // Queue delete stack if count reaches zero.
-            if (component.Count <= 0)
-                QueueDel(uid);
         }
 
         /// <summary>
         ///     Try to split this stack into two. Returns a non-null <see cref="Robust.Shared.GameObjects.EntityUid"/> if successful.
         /// </summary>
-        public EntityUid? Split(EntityUid uid, int amount, EntityCoordinates spawnPosition, StackComponent? stack = null)
+        public override EntityUid? Split(EntityUid uid, int amount, EntityCoordinates spawnPosition, StackComponent? stack = null)
         {
-            if (!Resolve(uid, ref stack))
-                return null;
+            EntityUid? entity = base.Split(uid, amount, spawnPosition, stack);
 
-            // Try to remove the amount of things we want to split from the original stack...
-            if (!Use(uid, amount, stack))
-                return null;
-
-            // Get a prototype ID to spawn the new entity. Null is also valid, although it should rarely be picked...
-            var prototype = _prototypeManager.TryIndex<StackPrototype>(stack.StackTypeId, out var stackType)
-                ? stackType.Spawn.ToString()
-                : Prototype(uid)?.ID;
-
-            // Set the output parameter in the event instance to the newly split stack.
-            var entity = Spawn(prototype, spawnPosition);
-
-            if (TryComp(entity, out StackComponent? stackComp))
-            {
-                // Set the split stack's count.
-                SetCount(entity, amount, stackComp);
-                // Don't let people dupe unlimited stacks
-                stackComp.Unlimited = false;
-            }
-
-            var ev = new StackSplitEvent(entity);
+            var ev = new StackSplitEvent(entity.GetValueOrDefault());
             RaiseLocalEvent(uid, ref ev);
 
             return entity;
@@ -165,63 +136,5 @@ namespace Content.Server.Stack
             return amounts;
         }
 
-        private void OnStackAlternativeInteract(EntityUid uid, StackComponent stack, GetVerbsEvent<AlternativeVerb> args)
-        {
-            if (!args.CanAccess || !args.CanInteract || args.Hands == null || stack.Count == 1)
-                return;
-
-            AlternativeVerb halve = new()
-            {
-                Text = Loc.GetString("comp-stack-split-halve"),
-                Category = VerbCategory.Split,
-                Act = () => UserSplit(uid, args.User, stack.Count / 2, stack),
-                Priority = 1
-            };
-            args.Verbs.Add(halve);
-
-            var priority = 0;
-            foreach (var amount in DefaultSplitAmounts)
-            {
-                if (amount >= stack.Count)
-                    continue;
-
-                AlternativeVerb verb = new()
-                {
-                    Text = amount.ToString(),
-                    Category = VerbCategory.Split,
-                    Act = () => UserSplit(uid, args.User, amount, stack),
-                    // we want to sort by size, not alphabetically by the verb text.
-                    Priority = priority
-                };
-
-                priority--;
-
-                args.Verbs.Add(verb);
-            }
-        }
-
-        private void UserSplit(EntityUid uid, EntityUid userUid, int amount,
-            StackComponent? stack = null,
-            TransformComponent? userTransform = null)
-        {
-            if (!Resolve(uid, ref stack))
-                return;
-
-            if (!Resolve(userUid, ref userTransform))
-                return;
-
-            if (amount <= 0)
-            {
-                Popup.PopupCursor(Loc.GetString("comp-stack-split-too-small"), userUid, PopupType.Medium);
-                return;
-            }
-
-            if (Split(uid, amount, userTransform.Coordinates, stack) is not {} split)
-                return;
-
-            Hands.PickupOrDrop(userUid, split);
-
-            Popup.PopupCursor(Loc.GetString("comp-stack-split"), userUid);
-        }
     }
 }
